+++ engine/src/modules/processing_graph/compiler/actions/copy_audio_buffer_action.h
@@ -20,6 +20,7 @@
 #pragma once
 
 #include <memory>
+#include <cstdint>
 
 #include <juce_core/juce_core.h>
 
@@ -32,16 +33,16 @@ private:
   JUCE_LEAK_DETECTOR(CopyAudioBufferAction)
 public:
   AnthemProcessContext* source;
-  int32_t sourcePortId;
+  int64_t sourcePortId;
 
   AnthemProcessContext* destination;
-  int32_t destinationPortId;
+  int64_t destinationPortId;
 
   CopyAudioBufferAction(
     AnthemProcessContext* source,
-    int32_t sourcePortId,
+    int64_t sourcePortId,
     AnthemProcessContext* destination,
-    int32_t destinationPortId
+    int64_t destinationPortId
   ) : source(source), sourcePortId(sourcePortId), destination(destination), destinationPortId(destinationPortId) {}
 
   void execute(int numSamples) override;
